Removed a lot of redndant cast after generisizing, but many remain.

+++ javatools/src/main/java/org/xvm/asm/constants/Float16Constant.java
@@ -74,7 +74,7 @@ public class Float16Constant
      */
     @Override
     public Float getValue() {
-        return Float.valueOf(m_flVal);
+        return m_flVal;
     }
